tests/functional/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.0
aiohttp==3.10.10
factory-boy==3.3.0
httpx==0.27.2
asyncpg==0.30.0
SQLAlchemy==2.0.36 #  db
Werkzeug==3.0.4 # checkout password
python-slugify
